[#50] Move auth.ts into the services directory
