+++ src/app.tsx
@@ -8,7 +8,7 @@ import {deleteCookie, getCookie, setCookie} from 'hono/cookie'
 import bcrypt from 'bcryptjs'
 import Layout from '@/layout'
 import account from '@/routes/account'
-import {AUTH_COOKIE_NAME, isLoggedIn} from '@/auth'
+import {AUTH_COOKIE_NAME, isLoggedIn} from '@/services/auth'
 import Input from '@/input'
 import panel from '@/routes/panel'
 import {db} from '@/db'
