+++ src/routes/account.tsx
@@ -2,7 +2,7 @@ import {Hono} from 'hono'
 import {getCookie} from 'hono/cookie'
 import Layout from '@/layout'
 import bcrypt from 'bcryptjs'
-import {AUTH_COOKIE_NAME, isLoggedIn} from '@/auth'
+import {AUTH_COOKIE_NAME, isLoggedIn} from '@/services/auth'
 import Input from '@/input'
 import {db} from '@/db'
 import {users} from '@/db/schema'
